Flatten Effect data structure into do_ methods. Those methods are the flattening of the debug/trace operation for each effect + the actual interpretation (which was previously in interpret_effect). Some of those methods can be further simplified improved, but for the sake of simplicity to review the PR I tried to keep this refactoring as mechanical as possible, without any business logic changes.
Over the next PRs I'll start moving some of the business logic within the rest of the state machine.
